+++ app/backend/src/ui/methods/projects/get-projects.ts
@@ -2,17 +2,10 @@ import Project from '../../../../../types/entities/project';
 import ProjectSchema from '../../../database/models/project';
 
 /**
- * @param workspaceId
+ * Returns list of stored projects
  */
-// eslint-disable-next-line no-unused-vars
-export async function getProjects(workspaceId: string | undefined): Promise<Project[]> {
-  let project;
-
-  await ProjectSchema.find()
-    .exec()
-    .then(result => {
-      project = result;
-    });
-
-  return project;
+export async function getProjects(): Promise<Project[]> {
+  return ProjectSchema
+    .find()
+    .exec();
 }
