--- app/backend/src/ui/methods/projects/update-picture.ts
@@ -1,19 +1,0 @@
-import Project from '../../../../../types/entities/project';
-import ProjectSchema from '../../../database/models/project';
-
-/**
- * @param id
- * @param newPicture
- */
-export async function updatePicture(id: string, newPicture: string): Promise<Project> {
-  let project;
-
-  await ProjectSchema.findByIdAndUpdate(id,
-    { picture: newPicture }, { new: true })
-    .exec()
-    .then(result => {
-      project = result;
-    });
-
-  return project;
-}
